Swapped placeholder object for dollar bills, and physics for interaction.
I added a collider and rigidbody to the bills, and got the feel of what it's like to interact with them in VR. With full interaction like this, I can handle about 100-120 bills really well. When it hits 200 though, things get a little shaky (at least on my PC). Started playing witha script to make the bills rotate/move as they fall, but it's still half-baked.